+++ workspace/metro/src.metro/metro/MetroThruTrack.java
@@ -28,11 +28,11 @@ public class MetroThruTrack extends MetroTrack {
         return outputPort;
     }
     @Override
-    public void processBuffer(Metro metro, int barLengthInFrames) throws MetroException {
+    public void processBuffer(Metro metro, long barLengthInFrames) throws MetroException {
     }
     @Override
     public void progressCursor(
-        Metro metro, int nframes, 
+        Metro metro, long nframes, 
         List<MetroMidiEvent> inputMidiEventList,
         List<MetroMidiEvent> outputMidiEventList) throws MetroException 
     {
